backend/requirements.txt
fastapi
uvicorn
pydantic
ollama
python-dotenv
PyJWT[crypto]
